--- src/app/auth-options/auth-options.page.scss
@@ -1,70 +1,0 @@
-.custom-header{
-    height: 50%;
-  }
-  
-.custom-toolbar{
-    height: 100%;
-    background-color: #774936;
-}
-  
-.custom-title{
-    text-align: right; 
-    text-decoration: underline;
-    font-size: 16px;
-    margin-top: 10px;
-}
-
-
-
-.bottom-link {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
-
-.image-container {
-    display: flex;
-    justify-content: center; 
-    align-items: center; 
-}
-
-.custom-image{
-    height: 300px;
-}
-  
-.custom-title2 {
-    color: #774936;
-    text-align: center;
-    margin-bottom: 20px;
-
-    a{
-        color: #774936;
-    }
-}
-
-  
-ion-button {
-    --background: #ffffff;
-    --background-hover: #ffffff;
-    --background-activated: #ffffff;
-    --background-focused: #ffffff;
-  
-    --color: #774936;
-  
-    --border-radius: 10px;
-    --border-color: #774936;
-    --border-style: solid;
-    --border-width: 1px;
-  
-    --box-shadow: 0 2px 6px 0 #774936;
-  
-    --ripple-color: #774936;
-  
-    --padding-top: 10px;
-    --padding-bottom: 10px;
-
-    
-}
