--- src/app/auth-sign-up/auth-sign-up.page.scss
@@ -1,81 +1,0 @@
-.formControl {
-    margin: 0 auto; 
-    margin-bottom: 20px;
-    width: 350px;
-}
-
-.custom-header{
-    height: 50%;
-  }
-  
-.custom-toolbar{
-    height: 100%;
-    background-color: #774936;
-}
-  
-.custom-title{
-    text-align: right; 
-    text-decoration: underline;
-    font-size: 16px;
-    margin-top: 10px;
-}
-
-.custom-title2 {
-    color: #774936;
-    text-align: center;
-    margin-bottom: 20px;
-
-    a{
-        color: #774936;
-    }
-}
-
-.image-container {
-    display: flex;
-    justify-content: center; 
-    align-items: center; 
-}
-
-.button-container1 {
-    display: flex;
-    align-items: center;
-}
-  
-
-.custom-image {
-    height: 300px;
-    margin-top: -40px; 
-}
-
-.custom-img{
-    height: 20px;
-    width: 20px;
-}
-
-.button-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-} 
-
-ion-button {
-    --background: #ffffff;
-    --background-hover: #ffffff;
-    --background-activated: #ffffff;
-    --background-focused: #ffffff;
-  
-    --color: #774936;
-  
-    --border-radius: 10px;
-    --border-color: #774936;
-    --border-style: solid;
-    --border-width: 1px;
-  
-    --box-shadow: 0 2px 6px 0 #774936;
-  
-    --ripple-color: #774936;
-  
-    --padding-top: 10px;
-    --padding-bottom: 10px;
-
-}
