+++ src/app/components/item-explorer/item-explorer.component.html
@@ -1,0 +1,10 @@
+<ion-card *ngFor="let product of products$ | async">
+  <img [src]="product.image" alt="Product image">
+  <ion-card-header>
+    <ion-card-title>{{ product.name }}</ion-card-title>
+    <ion-card-subtitle>${{ product.price }}</ion-card-subtitle>
+  </ion-card-header>
+  <ion-card-content>
+    Type: {{ product.type }}
+  </ion-card-content>
+</ion-card>
