+++ src/app/components/menu/menu.component.html
@@ -1,0 +1,31 @@
+<ion-menu contentId="main-content">
+  <ion-header>
+    <ion-toolbar>
+      <ion-title class="menu">Menu</ion-title>
+      <!-- List of menu items -->
+      <ion-list>
+        <ion-item>
+          <ion-label [routerLink]="['/home']">Home</ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label [routerLink]="['/auth']">Explore</ion-label>
+        </ion-item>
+        <ion-item>
+          <ion-label [routerLink]="['/profile']">Profile</ion-label>
+        </ion-item>
+      </ion-list>
+      <!-- List of menu items -->
+    </ion-toolbar>
+  </ion-header>
+</ion-menu>
+
+<div class="ion-page" id="main-content">
+  <ion-header>
+    <ion-toolbar>
+      <ion-buttons slot="start">
+        <ion-menu-button></ion-menu-button>
+      </ion-buttons>
+      <ion-title>Menu</ion-title>
+    </ion-toolbar>
+  </ion-header>
+</div>
