--- src/app/home/home.page.html
@@ -1,6 +1,0 @@
-<ion-header>
-  <ion-toolbar>
-      
-  </ion-toolbar>
-</ion-header>
-
