+++ src/app/pages/home/home.page.html
@@ -1,0 +1,6 @@
+<ion-content [fullscreen]="true">
+  <app-menu></app-menu>
+  <div class="container" id="container">
+    <app-item-explorer></app-item-explorer>
+  </div>
+</ion-content>
