+++ src/app/pages/home/home.page.scss
@@ -1,0 +1,3 @@
+.container{
+    margin-top: 60px;
+}
