+++ src/app/pages/profile/profile.page.html
@@ -1,24 +1,15 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-menu-button></ion-menu-button>
-    </ion-buttons>
-    <ion-title>Profile</ion-title>
-  </ion-toolbar>
-</ion-header>
-
 <ion-content>
+  <app-menu class="menu"></app-menu>
   <ion-header collapse="condense">
     <ion-toolbar>
     </ion-toolbar>
-
   </ion-header>
   <ion-card class="user-description-card">
     <ion-card-content>
       <ion-list lines="none">
         <ion-item>
           <ion-avatar slot="start">
-            <img src="/assets/icon/profile.jpg"/>
+            <img src="https://github.com/pixelforgeutp.png"/>
           </ion-avatar>
           <ion-label>
             <h1>Lee Sang-hyeok</h1>
